refactor: 💡 move dataStore instance to AppSettings.kt
